[#60] Remove equivalency system
This was a system that added a lot of complexity for the sole benefit of
saving a number of bytes in the list capabilities response. In the end,
the benefit didn't outweigh the costs in my opinion.

If the cost of storing the response on the board is really too much, you
have a Turing-complete language at your disposal to try and save bytes
through using the repetition.