[#61] Remove defining allowed data values
This was once again a lot of complexity for little gain. Removing this
made the list capabilities response much more grokkable while the
contrivedness of the examples showed that it's not all that likely that
boards will actually only support some data and not other data.

On top of that, since the response length has been removed from the
protocol, this was also the last occurence of the length system, which
can now also be removed for further simplification.